chore: rename hook file to match hook name
